worker/requirements.txt
pika>=1.3.0
supabase>=2.0.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
tenacity>=8.2.0
